move decoy api into separate module
