+++ src/web/mod.rs
@@ -1,4 +1,2 @@
 /// Endpoint for dealing with decoys
-pub mod decoy_controller;
-/// Web server definition and configuration
-pub mod server;
+pub mod decoy_api;
